feat: Generate invoices in CSV format
This commit adds functionality to generate invoices in CSV format. The `execute` method in the `GenerateInvoices` class now checks if the `format` property in the input is set to "csv". If so, it generates the invoices in CSV format by converting the output array into a string of lines joined by semicolons and newlines. This allows users to easily export the generated invoices in a CSV file format.